post api 형식 검증 로직 추가
