+++ src/app/(pages)/post/edit/[postId]/page.tsx
@@ -1,6 +1,8 @@
-import { redirect } from "next/navigation";
+import { forbidden, redirect, unauthorized } from "next/navigation";
 
 import { getPost } from "@/entities/post";
+import { getSession } from "@/features/request-sign-in";
+import { Session } from "@/shared/types/auth-types";
 import { Post } from "@/shared/types/post-types";
 import EditPost from "@/widgets/edit-post";
 
@@ -15,10 +17,16 @@ export default async function EditPostPage({
     post = await getPost(postId);
   }
 
+  const session: Session | null = await getSession();
+
+  if (!session) {
+    forbidden();
+  }
+
   if (!post) {
     console.log("no post");
     redirect("/post/new");
   } else {
-    return <EditPost post={post} />;
+    return <EditPost post={post} session={session} />;
   }
 }
