+++ src/app/(pages)/post/new/page.tsx
@@ -1,5 +1,12 @@
+import { forbidden } from "next/navigation";
+
+import { getSession } from "@/features/request-sign-in";
 import EditPost from "@/widgets/edit-post";
 
-export default function NewPostPage() {
-  return <EditPost />;
+export default async function NewPostPage() {
+  const session = await getSession();
+  if (!session) {
+    forbidden();
+  }
+  return <EditPost session={session} />;
 }
