+++ src/components/editor/use-create-editor.ts
@@ -90,66 +90,13 @@ import { MentionInputElement } from "@/components/plate-ui/mention-input-element
 import { ParagraphElement } from "@/components/plate-ui/paragraph-element";
 import { withPlaceholders } from "@/components/plate-ui/placeholder";
 import { SlashInputElement } from "@/components/plate-ui/slash-input-element";
-import {
-  TableCellElement,
-  TableCellHeaderElement,
-} from "@/components/plate-ui/table-cell-element";
-import { TableElement } from "@/components/plate-ui/table-element";
-import { TableRowElement } from "@/components/plate-ui/table-row-element";
 import { TocElement } from "@/components/plate-ui/toc-element";
 import { ToggleElement } from "@/components/plate-ui/toggle-element";
 
-export const useCreateEditor = () => {
+export const useCreateEditor = (value: string) => {
   return usePlateEditor({
     override: {
-      components: withPlaceholders({
-        // [AIPlugin.key]: AILeaf,
-        // [AudioPlugin.key]: MediaAudioElement,
-        [BlockquotePlugin.key]: BlockquoteElement,
-        [BoldPlugin.key]: withProps(PlateLeaf, { as: "strong" }),
-        // [CodeBlockPlugin.key]: CodeBlockElement,
-        [CodeLinePlugin.key]: CodeLineElement,
-        // [CodePlugin.key]: CodeLeaf,
-        // [CodeSyntaxPlugin.key]: CodeSyntaxLeaf,
-        [ColumnItemPlugin.key]: ColumnElement,
-        [ColumnPlugin.key]: ColumnGroupElement,
-        [CommentsPlugin.key]: CommentLeaf,
-        [DatePlugin.key]: DateElement,
-        // [EmojiInputPlugin.key]: EmojiInputElement,
-        // [EquationPlugin.key]: EquationElement,
-        [ExcalidrawPlugin.key]: ExcalidrawElement,
-        [FilePlugin.key]: MediaFileElement,
-        [HEADING_KEYS.h1]: withProps(HeadingElement, { variant: "h1" }),
-        [HEADING_KEYS.h2]: withProps(HeadingElement, { variant: "h2" }),
-        [HEADING_KEYS.h3]: withProps(HeadingElement, { variant: "h3" }),
-        [HEADING_KEYS.h4]: withProps(HeadingElement, { variant: "h4" }),
-        [HEADING_KEYS.h5]: withProps(HeadingElement, { variant: "h5" }),
-        [HEADING_KEYS.h6]: withProps(HeadingElement, { variant: "h6" }),
-        [HighlightPlugin.key]: HighlightLeaf,
-        [HorizontalRulePlugin.key]: HrElement,
-        [ImagePlugin.key]: ImageElement,
-        [InlineEquationPlugin.key]: InlineEquationElement,
-        [ItalicPlugin.key]: withProps(PlateLeaf, { as: "em" }),
-        [KbdPlugin.key]: KbdLeaf,
-        [LinkPlugin.key]: LinkElement,
-        [MediaEmbedPlugin.key]: MediaEmbedElement,
-        [MentionInputPlugin.key]: MentionInputElement,
-        [MentionPlugin.key]: MentionElement,
-        [ParagraphPlugin.key]: ParagraphElement,
-        [PlaceholderPlugin.key]: MediaPlaceholderElement,
-        [SlashInputPlugin.key]: SlashInputElement,
-        [StrikethroughPlugin.key]: withProps(PlateLeaf, { as: "s" }),
-        [SubscriptPlugin.key]: withProps(PlateLeaf, { as: "sub" }),
-        [SuperscriptPlugin.key]: withProps(PlateLeaf, { as: "sup" }),
-        // [TableCellHeaderPlugin.key]: TableCellHeaderElement,
-        // [TableCellPlugin.key]: TableCellElement,
-        // [TablePlugin.key]: TableElement,
-        // [TableRowPlugin.key]: TableRowElement,
-        [TocPlugin.key]: TocElement,
-        [TogglePlugin.key]: ToggleElement,
-        [UnderlinePlugin.key]: withProps(PlateLeaf, { as: "u" }),
-        // [VideoPlugin.key]: MediaVideoElement,
-      }),
+      components: withPlaceholders(components),
     },
     plugins: [
       // ...copilotPlugins,
@@ -157,6 +104,55 @@ export const useCreateEditor = () => {
       FixedToolbarPlugin,
       FloatingToolbarPlugin,
     ],
-    value: [],
+    value: value,
   });
+};
+
+export const components = {
+  // [AIPlugin.key]: AILeaf,
+  // [AudioPlugin.key]: MediaAudioElement,
+  [BlockquotePlugin.key]: BlockquoteElement,
+  [BoldPlugin.key]: withProps(PlateLeaf, { as: "strong" }),
+  // [CodeBlockPlugin.key]: CodeBlockElement,
+  [CodeLinePlugin.key]: CodeLineElement,
+  // [CodePlugin.key]: CodeLeaf,
+  // [CodeSyntaxPlugin.key]: CodeSyntaxLeaf,
+  [ColumnItemPlugin.key]: ColumnElement,
+  [ColumnPlugin.key]: ColumnGroupElement,
+  [CommentsPlugin.key]: CommentLeaf,
+  [DatePlugin.key]: DateElement,
+  // [EmojiInputPlugin.key]: EmojiInputElement,
+  // [EquationPlugin.key]: EquationElement,
+  [ExcalidrawPlugin.key]: ExcalidrawElement,
+  [FilePlugin.key]: MediaFileElement,
+  [HEADING_KEYS.h1]: withProps(HeadingElement, { variant: "h1" }),
+  [HEADING_KEYS.h2]: withProps(HeadingElement, { variant: "h2" }),
+  [HEADING_KEYS.h3]: withProps(HeadingElement, { variant: "h3" }),
+  [HEADING_KEYS.h4]: withProps(HeadingElement, { variant: "h4" }),
+  [HEADING_KEYS.h5]: withProps(HeadingElement, { variant: "h5" }),
+  [HEADING_KEYS.h6]: withProps(HeadingElement, { variant: "h6" }),
+  [HighlightPlugin.key]: HighlightLeaf,
+  [HorizontalRulePlugin.key]: HrElement,
+  [ImagePlugin.key]: ImageElement,
+  [InlineEquationPlugin.key]: InlineEquationElement,
+  [ItalicPlugin.key]: withProps(PlateLeaf, { as: "em" }),
+  [KbdPlugin.key]: KbdLeaf,
+  [LinkPlugin.key]: LinkElement,
+  [MediaEmbedPlugin.key]: MediaEmbedElement,
+  [MentionInputPlugin.key]: MentionInputElement,
+  [MentionPlugin.key]: MentionElement,
+  [ParagraphPlugin.key]: ParagraphElement,
+  [PlaceholderPlugin.key]: MediaPlaceholderElement,
+  [SlashInputPlugin.key]: SlashInputElement,
+  [StrikethroughPlugin.key]: withProps(PlateLeaf, { as: "s" }),
+  [SubscriptPlugin.key]: withProps(PlateLeaf, { as: "sub" }),
+  [SuperscriptPlugin.key]: withProps(PlateLeaf, { as: "sup" }),
+  // [TableCellHeaderPlugin.key]: TableCellHeaderElement,
+  // [TableCellPlugin.key]: TableCellElement,
+  // [TablePlugin.key]: TableElement,
+  // [TableRowPlugin.key]: TableRowElement,
+  [TocPlugin.key]: TocElement,
+  [TogglePlugin.key]: ToggleElement,
+  [UnderlinePlugin.key]: withProps(PlateLeaf, { as: "u" }),
+  // [VideoPlugin.key]: MediaVideoElement,
 };
